Cleaning, Optimizing and Adding DoubleTap Gesture Event
Cleaning up the code, taking out all of the debugging stuff that is no longer needed. Tried to optimize things a bit by checking if an event is registered before even bothering to check its state. So if nothing registers to a trigger event, dont waste time checking and testing for it. Also began playing with creating gestures with the touchpad, so far I put together  a somple Double Tap gesture and event to the manager.